requirements.txt

streamlit>=1.28.0
rembg>=2.0.50
Pillow>=9.5.0
numpy>=1.24.0
onnxruntime>=1.15.0
opencv-python>=4.8.0